+++ sdk-lib/src/main/java/cash/z/ecc/android/sdk/model/TransactionOverview.kt
@@ -12,20 +12,19 @@ import cash.z.ecc.android.sdk.internal.model.DbTransactionOverview
  * last synced block exceeds the [expiryHeight].
  */
 data class TransactionOverview internal constructor(
-    val id: Long,
     val rawId: FirstClassByteArray,
     val minedHeight: BlockHeight?,
     val expiryHeight: BlockHeight?,
-    val index: Long,
+    val index: Long?,
     val raw: FirstClassByteArray?,
     val isSentTransaction: Boolean,
     val netValue: Zatoshi,
-    val feePaid: Zatoshi,
+    val feePaid: Zatoshi?,
     val isChange: Boolean,
     val receivedNoteCount: Int,
     val sentNoteCount: Int,
     val memoCount: Int,
-    val blockTimeEpochSeconds: Long,
+    val blockTimeEpochSeconds: Long?,
     val transactionState: TransactionState
 ) {
     override fun toString() = "TransactionOverview"
@@ -36,7 +35,6 @@ data class TransactionOverview internal constructor(
             latestBlockHeight: BlockHeight?
         ): TransactionOverview {
             return TransactionOverview(
-                dbTransactionOverview.id,
                 dbTransactionOverview.rawId,
                 dbTransactionOverview.minedHeight,
                 dbTransactionOverview.expiryHeight,
@@ -73,17 +71,29 @@ enum class TransactionState {
             minedHeight: BlockHeight?,
             expiryHeight: BlockHeight?
         ): TransactionState {
-            return if (latestBlockHeight == null) {
-                Pending
-            } else if (minedHeight != null && (latestBlockHeight.value - minedHeight.value) >= MIN_CONFIRMATIONS) {
-                Confirmed
-            } else if (minedHeight != null && (latestBlockHeight.value - minedHeight.value) < MIN_CONFIRMATIONS) {
-                Pending
-            } else if (minedHeight == null && ((expiryHeight?.value ?: Long.MAX_VALUE) < latestBlockHeight.value)) {
-                Pending
-            } else {
-                Expired
-            }
+            return latestBlockHeight?.let { chainTip ->
+                minedHeight?.let { minedHeight ->
+                    // A transaction mined in the latest block has 1 confirmation.
+                    if ((chainTip + 1 - minedHeight) >= MIN_CONFIRMATIONS) {
+                        Confirmed
+                    } else {
+                        Pending
+                    }
+                } ?: expiryHeight?.let { expiryHeight ->
+                    // Expiry height is the last height at which a transaction can be mined.
+                    // If the chain tip is greater than or equal to the expiry height, the
+                    // transaction can never be mined. A value of 0 disables expiry.
+                    if (expiryHeight.value == 0L || expiryHeight > chainTip) {
+                        Pending
+                    } else {
+                        Expired
+                    }
+                }
+                // Base case: either we don't know the latest block height (unlikely if we
+                // know about transactions), or the transaction is both unmined and has an
+                // unknown expiry height (because we haven't seen the full transaction).
+                // Treat these as Pending because the status will change as we sync.
+            } ?: Pending
         }
     }
 }